Fix static

+++ mindspore/core/ops/rpc_send.cc
@@ -28,7 +28,7 @@ AbstractBasePtr RpcSendInfer(const abstract::AnalysisEnginePtr &, const Primitiv
   if (input_args.empty()) {
     MS_LOG(EXCEPTION) << "The input size of RpcSend is 0.";
   }
-  if (input_args.size() == kDim1) {
+  if (input_args.size() == static_cast<size_t>(kDim1)) {
     return input_args[kInputIndex0];
   } else {
     abstract::AbstractTuplePtr rpc_send_abs = std::make_shared<abstract::AbstractTuple>(input_args);
